+++ src/main/java/www/egg/hom/HomeController.java
@@ -1,8 +1,6 @@
 package www.egg.hom;
 
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
@@ -10,11 +8,6 @@ import org.springframework.web.bind.annotation.RequestMethod;
 //@Controller
 public class HomeController {
 
-	@RequestMapping(value = "why", method = RequestMethod.GET)
-	public String why(Model model) {
-		
-		return "ownPage";
-	}
 
 	@RequestMapping(value = "/", method = RequestMethod.GET)
 	public String home(Model model) {
